worker/requirements.txt
# Python dependencies required by your worker
requests
psycopg2-binary
